test: Add tests for expression helpers.
